initial attempt at a plugin system
- currently only covers commands
- some commands which need to maintain state such as seen might not work properly yet!
